--- src/Controllers/ContentController.hh
@@ -1,15 +1,0 @@
-<?hh //strict
-
-      namespace HelloWorld\Controllers;
-
-
-      use Plenty\Plugin\Controller;
-      use Plenty\Plugin\Templates\Twig;
-
-      class ContentController extends Controller
-      {
-    	    public function sayHello(Twig $twig):string
-    	    {
-        		  return $twig->render('HelloWorld::content.hello');
-    	    }
-      }
