--- src/Providers/HelloWorldRouteServiceProvider.hh
@@ -1,16 +1,0 @@
-<?hh //strict
-
-      namespace HelloWorld\Providers;
-
-
-      use Plenty\Plugin\RouteServiceProvider;
-      use Plenty\Plugin\Routing\Router;
-
-      class HelloWorldRouteServiceProvider extends RouteServiceProvider
-      {
-        public function map(Router $router):void
-        {
-          $router->get('hello', 'HelloWorld\Controllers\ContentController@sayHello');
-        }
-
-      }
