+++ src/Providers/HelloWorldRouteServiceProvider.php
@@ -1,0 +1,21 @@
+<?php
+namespace HelloWorld\Providers;
+
+use Plenty\Plugin\RouteServiceProvider;
+use Plenty\Plugin\Routing\Router;
+
+/**
+ * Class HelloWorldRouteServiceProvider
+ * @package HelloWorld\Providers
+ */
+class HelloWorldRouteServiceProvider extends RouteServiceProvider
+{
+	/**
+	 * @param Router $router
+	 */
+	public function map(Router $router):void
+	{
+		$router->get('hello', 'HelloWorld\Controllers\ContentController@sayHello');
+	}
+
+}
